{{cookiecutter.project_slug}}/requirements.txt
## Python requirements for using the {{ cookiecutter.app_name }} application.
##
##   pyhon -m pip install --requirement=requirements.txt
##
python-dotenv>=0.12
PyYAML>=3.11
